+++ src/containerapp/azext_containerapp/tests/latest/test_containerappjob_preview_scenario.py
@@ -4,50 +4,24 @@
 # --------------------------------------------------------------------------------------------
 
 import os
-import time
-from time import sleep
 from azure.cli.testsdk import (ScenarioTest, ResourceGroupPreparer, JMESPathCheck, live_only)
-from subprocess import run
 
-from .common import (write_test_file, TEST_LOCATION, clean_up_test_file)
+from azure.cli.testsdk.decorators import serial_test
+from .common import TEST_LOCATION
 from .custom_preparers import ConnectedClusterPreparer
-from .utils import create_containerapp_env
+from .utils import create_containerapp_env, create_extension_and_custom_location
 
 TEST_DIR = os.path.abspath(os.path.join(os.path.abspath(__file__), '..'))
 
 
 class ContainerappJobPreviewScenarioTest(ScenarioTest):
-
+    @serial_test()
     @ResourceGroupPreparer(location=TEST_LOCATION, random_name_length=15)
     @ConnectedClusterPreparer(location=TEST_LOCATION)
     def test_containerappjob_preview_environment_type(self, resource_group, infra_cluster, connected_cluster_name):
         self.cmd('configure --defaults location={}'.format(TEST_LOCATION))
         custom_location_name = "my-custom-location"
-        try:
-            connected_cluster = self.cmd(f'az connectedk8s show --resource-group {resource_group} --name {connected_cluster_name}').get_output_in_json()
-            while connected_cluster["connectivityStatus"] == "Connecting":
-                time.sleep(5)
-                connected_cluster = self.cmd(f'az connectedk8s show --resource-group {resource_group} --name {connected_cluster_name}').get_output_in_json()
-
-            connected_cluster_id = connected_cluster.get('id')
-            extension = self.cmd(f'az k8s-extension create'
-                                 f' --resource-group {resource_group}'
-                                 f' --name containerapp-ext'
-                                 f' --cluster-type connectedClusters'
-                                 f' --cluster-name {connected_cluster_name}'
-                                 f' --extension-type "Microsoft.App.Environment" '
-                                 f' --release-train stable'
-                                 f' --auto-upgrade-minor-version true'
-                                 f' --scope cluster'
-                                 f' --release-namespace appplat-ns'
-                                 f' --configuration-settings "Microsoft.CustomLocation.ServiceAccount=default"'
-                                 f' --configuration-settings "appsNamespace=appplat-ns"'
-                                 f' --configuration-settings "clusterName={connected_cluster_name}"'
-                                 f' --configuration-settings "envoy.annotations.service.beta.kubernetes.io/azure-load-balancer-resource-group={resource_group}"').get_output_in_json()
-
-            self.cmd(f'az customlocation create -g {resource_group} -n {custom_location_name} -l {TEST_LOCATION} --host-resource-id {connected_cluster_id} --namespace appplat-ns -c {extension["id"]}')
-        except:
-            pass
+        create_extension_and_custom_location(self, resource_group, connected_cluster_name, custom_location_name)
 
         # create connected environment with client or create a command for connected?
         sub_id = self.cmd('az account show').get_output_in_json()['id']
@@ -55,16 +29,11 @@ class ContainerappJobPreviewScenarioTest(ScenarioTest):
         connected_env_name = 'my-connected-env'
         connected_env_resource_id = f"/subscriptions/{sub_id}/resourceGroups/{resource_group}/providers/Microsoft.App/connectedEnvironments/{connected_env_name}"
         custom_location_id = f"/subscriptions/{sub_id}/resourceGroups/{resource_group}/providers/Microsoft.ExtendedLocation/customLocations/{custom_location_name}"
-        file = f"{resource_group}.json"
-        env_payload = '{{ "location": "{location}", "extendedLocation": {{ "name": "{custom_location_id}", "type": "CustomLocation" }}, "Properties": {{}}}}' \
-            .format(location=TEST_LOCATION, custom_location_id=custom_location_id)
-        write_test_file(file, env_payload)
-        self.cmd(f'az rest --method put --uri "{connected_env_resource_id}?api-version=2022-06-01-preview" --body "@{file}"')
-        containerapp_env = self.cmd(f'az rest --method get --uri "{connected_env_resource_id}?api-version=2022-06-01-preview"').get_output_in_json()
-        while containerapp_env["properties"]["provisioningState"].lower() != "succeeded":
-            time.sleep(5)
-            containerapp_env = self.cmd(
-                f'az rest --method get --uri "{connected_env_resource_id}?api-version=2022-06-01-preview"').get_output_in_json()
+        self.cmd(f'containerapp connected-env create -g {resource_group} --name {connected_env_name} --custom-location {custom_location_name} -l {TEST_LOCATION}', checks=[
+               JMESPathCheck('name', connected_env_name),
+               JMESPathCheck('properties.provisioningState', "Succeeded"),
+               JMESPathCheck('extendedLocation.name', custom_location_id)
+        ])
 
         ca_name = self.create_random_name(prefix='containerapp', length=24)
         self.cmd(
@@ -104,7 +73,6 @@ class ContainerappJobPreviewScenarioTest(ScenarioTest):
         self.cmd('containerapp job list -g {}'.format(resource_group), checks=[
             JMESPathCheck('length(@)', 0)
         ])
-        clean_up_test_file(file)
 
     @ResourceGroupPreparer(location="eastus")
     def test_containerappjob_preview_e2e(self, resource_group):