+++ src/aks-preview/HISTORY.rst
@@ -12,10 +12,18 @@ To release a new version, please select a new version number (usually plus 1 to 
 Pending
 +++++++
 
-0.5.158
+0.5.160
 +++++++
 * Support `premium` cluster sku tier in `az aks create` and `az aks update` commands
 * Add option `--k8s-support-plan` to `az aks create` and `az aks update` commands
+
+0.5.159
++++++++
+* Revert `az aks copilot` Command
+
+0.5.158
++++++++
+* Add `enable-egress-gateway` and `disable-egress-gateway` to the `az aks mesh` command.
 
 0.5.157
 +++++++